+++ R/etox.R
@@ -4,7 +4,6 @@
 #' Targets \url{https://webetox.uba.de/webETOX/index.do} for their substance ID
 #'
 #' @import xml2 httr
-#' @importFrom stats rgamma
 #' @importFrom dplyr bind_rows
 #' @importFrom tibble tibble
 #' @param query character; The searchterm
@@ -26,7 +25,7 @@
 #' @references Eduard Szöcs, Tamás Stirling, Eric R. Scott, Andreas Scharmüller,
 #' Ralf B. Schäfer (2020). webchem: An R Package to Retrieve Chemical
 #' Information from the Web. Journal of Statistical Software, 93(13).
-#' <doi:10.18637/jss.v093.i13>.
+#' \doi{10.18637/jss.v093.i13}.
 #' @export
 #' @examples
 #' \dontrun{
@@ -83,7 +82,7 @@ get_etoxid <- function(query,
                    'stoffnummer.selection[0].type' = type,
                    event = "Search")
     }
-    Sys.sleep(stats::rgamma(1, shape = 15, scale = 1/10))
+    webchem_sleep(type = 'scrape')
     h <- try(httr::RETRY("POST",
                          url = baseurl,
                          httr::user_agent(webchem_url()),
@@ -138,7 +137,6 @@ get_etoxid <- function(query,
 #'
 #' @import xml2
 #' @importFrom rvest html_table
-#' @importFrom stats rgamma
 #' @param id character; ETOX ID
 #' @param verbose logical; print message during processing to console?
 #'
@@ -155,7 +153,7 @@ get_etoxid <- function(query,
 #' @references Eduard Szöcs, Tamás Stirling, Eric R. Scott, Andreas Scharmüller,
 #' Ralf B. Schäfer (2020). webchem: An R Package to Retrieve Chemical
 #' Information from the Web. Journal of Statistical Software, 93(13).
-#' <doi:10.18637/jss.v093.i13>.
+#' \doi{10.18637/jss.v093.i13}.
 #' @export
 #' @examples
 #' \dontrun{
@@ -182,7 +180,7 @@ etox_basic <- function(id, verbose = TRUE) {
     baseurl <- 'https://webetox.uba.de/webETOX/public/basics/stoff.do?language=en&id='
     qurl <- paste0(baseurl, id)
     if(verbose) webchem_message("query", id, appendLF = FALSE)
-    Sys.sleep(stats::rgamma(1, shape = 15, scale = 1 / 10))
+    webchem_sleep(type = 'scrape')
     res <- try(httr::RETRY("GET",
                            qurl,
                            httr::user_agent(webchem_url()),
@@ -253,7 +251,7 @@ etox_basic <- function(id, verbose = TRUE) {
     }
     }
   out <- lapply(id, foo, verbose = verbose)
-  out <- setNames(out, id)
+  names(out) <- id
   class(out) <- c('etox_basic', 'list')
   return(out)
 }
@@ -264,9 +262,8 @@ etox_basic <- function(id, verbose = TRUE) {
 #' Query ETOX: Information System Ecotoxicology and Environmental Quality
 #' Targets \url{https://webetox.uba.de/webETOX/index.do} for quality targets
 #'
-#' @import xml2 RCurl
+#' @import xml2
 #' @importFrom utils read.table
-#' @importFrom stats rgamma
 #' @param id character; ETOX ID
 #' @param verbose logical; print message during processing to console?
 #'
@@ -281,7 +278,7 @@ etox_basic <- function(id, verbose = TRUE) {
 #' @references Eduard Szöcs, Tamás Stirling, Eric R. Scott, Andreas Scharmüller,
 #' Ralf B. Schäfer (2020). webchem: An R Package to Retrieve Chemical
 #' Information from the Web. Journal of Statistical Software, 93(13).
-#' <doi:10.18637/jss.v093.i13>.
+#' \doi{10.18637/jss.v093.i13}.
 #' @export
 #' @examples
 #' \dontrun{
@@ -304,7 +301,7 @@ etox_targets <- function(id, verbose = TRUE) {
     baseurl <- 'https://webetox.uba.de/webETOX/public/basics/stoff.do?language=en&id='
     qurl <- paste0(baseurl, id)
     if(verbose) webchem_message("query", id, appendLF = FALSE)
-    Sys.sleep(stats::rgamma(1, shape = 15, scale = 1/10))
+    webchem_sleep(type = 'scrape')
     res <- try(httr::RETRY("GET",
                            qurl,
                            httr::user_agent(webchem_url()),
@@ -358,7 +355,7 @@ etox_targets <- function(id, verbose = TRUE) {
     }
   }
   out <- lapply(id, foo, verbose = verbose)
-  out <- setNames(out, id)
+  names(out) <- id
   return(out)
 }
 
@@ -368,9 +365,8 @@ etox_targets <- function(id, verbose = TRUE) {
 #' Query ETOX: Information System Ecotoxicology and Environmental Quality Targets
 #' \url{https://webetox.uba.de/webETOX/index.do} for tests
 #'
-#' @import xml2 RCurl
+#' @import xml2
 #' @importFrom utils read.table
-#' @importFrom stats rgamma
 #' @param id character; ETOX ID
 #' @param verbose logical; print message during processing to console?
 #'
@@ -402,7 +398,7 @@ etox_tests <- function(id, verbose = TRUE) {
     baseurl <- 'https://webetox.uba.de/webETOX/public/basics/stoff.do?id='
     qurl <- paste0(baseurl, id)
     if(verbose) webchem_message("query", id, appendLF = FALSE)
-    Sys.sleep(stats::rgamma(1, shape = 15, scale = 1/10))
+    webchem_sleep(type = 'scrape')
     res <- try(httr::RETRY("GET",
                            qurl,
                            httr::user_agent(webchem_url()),
@@ -451,6 +447,6 @@ etox_tests <- function(id, verbose = TRUE) {
     }
   }
   out <- lapply(id, foo, verbose = verbose)
-  out <- setNames(out, id)
+  names(out) <- id
   return(out)
 }